NSO/6.2.a_Module6_Lab_01/requirements.txt
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
executing==0.8.3
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
PyYAML==6.0
six==1.16.0
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
